requirements.txt
#google-generativeai==0.7.2
streamlit==1.35.0
python-dotenv==1.0.1
pandas==2.2.2
plotly==5.22.0
pydantic==2.7.4
httpx==0.27.0
tenacity==8.3.0
google-genai